Split up base QIS and helios QIS, allowing for Utilities to bind to a base rather than to something platform specific. Move to dynamic libs rather than static libs to achieve support for multiple utilities
