+++ pyproject.toml
@@ -77,6 +77,7 @@ packages = [
   "selene-ext/error-models/depolarizing/python/selene_depolarizing_error_model_plugin",
   "selene-ext/error-models/ideal/python/selene_ideal_error_model_plugin",
   "selene-ext/error-models/simple-leakage/python/selene_simple_leakage_error_model_plugin",
+  "selene-ext/interfaces/base_qis/python/selene_base_qis_plugin",
   "selene-ext/interfaces/helios_qis/python/selene_helios_qis_plugin",
   "selene-ext/runtimes/simple/python/selene_simple_runtime_plugin",
   "selene-ext/runtimes/soft_rz/python/selene_soft_rz_runtime_plugin",
@@ -105,6 +106,7 @@ cache-keys = [
   { file = "selene-ext/*/*/c_interface/**/*.h" },
   { file = "selene-ext/*/*/c_interface/**/CMakeLists.txt" },
   { file = "selene-ext/*/*/rust/**/*.rs" },
+  { file = "selene-ext/**/build.rs" },
   { file = "selene-ext/**/Cargo.toml" },
   { file = "selene-ext/**/Cargo.lock" },
 ]
@@ -133,6 +135,7 @@ known_first_party = [
   "selene_ideal_error_model_plugin",
   "selene_simple_leakage_error_model_plugin",
   # Interfaces
+  "selene_base_qis_plugin",
   "selene_helios_qis_plugin",
   # Runtimes
   "selene_simple_runtime_plugin",
