+++ selene-ext/interfaces/base_qis/c/.gitignore
@@ -1,0 +1,1 @@
+build/
