+++ selene-ext/interfaces/base_qis/c/CMakeLists.txt
@@ -1,21 +1,22 @@
 cmake_minimum_required(VERSION 3.12)
 
-project(HeliosSeleneInterface LANGUAGES C)
+project(BaseQISSeleneInterface LANGUAGES C)
 
 include(GNUInstallDirs)
 include(CMakePackageConfigHelpers)
 
 find_package(Selene REQUIRED)
-find_package(BaseQISSeleneInterface REQUIRED)
 get_target_property(selene_include_dirs Selene::Selene INTERFACE_INCLUDE_DIRECTORIES)
 
-set(base_name "helios_selene_interface")
+set(base_name "base_qis_selene_interface")
 
-function(build_helios_interface log_level suffix)
+file(GLOB_RECURSE source_files src/*.c)
 
-    set(runtime_target "${base_name}_runtime${suffix}")
+function(build_base_qis_interface log_level suffix)
 
-    add_library(${runtime_target} SHARED src/interface.c src/helios_ops.c src/main.c)
+    set(runtime_target "${base_name}${suffix}")
+
+    add_library(${runtime_target} SHARED ${source_files})
     set_target_properties(${runtime_target} PROPERTIES
         POSITION_INDEPENDENT_CODE ON
         C_STANDARD 11
@@ -27,13 +28,12 @@ function(build_helios_interface log_level suffix)
         "$<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>"
     )
     target_link_libraries(${runtime_target} PRIVATE Selene::Selene)
-    target_link_libraries(${runtime_target} PUBLIC BaseQISSeleneInterface::base_qis_selene_interface${suffix})
     if(MSVC)
         target_compile_options(${runtime_target} PRIVATE /TC /Zl)
     endif()
 
     install(TARGETS ${runtime_target}
-            EXPORT HeliosSeleneInterface
+            EXPORT BaseQISSeleneInterface
             RUNTIME DESTINATION ${CMAKE_INSTALL_LIBDIR}
             LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
             ARCHIVE DESTINATION ${CMAKE_INSTALL_LIBDIR}
@@ -42,27 +42,27 @@ function(build_helios_interface log_level suffix)
 
 endfunction()
 
-build_helios_interface(0 "")
-build_helios_interface(1 "_debug")
-build_helios_interface(2 "_diagnostic")
+build_base_qis_interface(0 "")
+build_base_qis_interface(1 "_debug")
+build_base_qis_interface(2 "_diagnostic")
 
-install(DIRECTORY include/helios_qis
+install(DIRECTORY include/base_qis
         DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}
 )
 install(
-    EXPORT HeliosSeleneInterface
-    FILE HeliosSeleneInterfaceTargets.cmake
-    NAMESPACE HeliosSeleneInterface::
-    DESTINATION ${CMAKE_INSTALL_LIBDIR}/cmake/HeliosSeleneInterface
+    EXPORT BaseQISSeleneInterface
+    FILE BaseQISSeleneInterfaceTargets.cmake
+    NAMESPACE BaseQISSeleneInterface::
+    DESTINATION ${CMAKE_INSTALL_LIBDIR}/cmake/BaseQISSeleneInterface
 )
 configure_package_config_file(
-    "${CMAKE_CURRENT_SOURCE_DIR}/cmake/HeliosSeleneInterfaceConfig.cmake.in"
-    "${CMAKE_CURRENT_BINARY_DIR}/HeliosSeleneInterfaceConfig.cmake"
-    INSTALL_DESTINATION ${CMAKE_INSTALL_LIBDIR}/cmake/HeliosSeleneInterface
+    "${CMAKE_CURRENT_SOURCE_DIR}/cmake/BaseQISSeleneInterfaceConfig.cmake.in"
+    "${CMAKE_CURRENT_BINARY_DIR}/BaseQISSeleneInterfaceConfig.cmake"
+    INSTALL_DESTINATION ${CMAKE_INSTALL_LIBDIR}/cmake/BaseQISSeleneInterface
 )
 install(
     FILES
-        "${CMAKE_CURRENT_BINARY_DIR}/HeliosSeleneInterfaceConfig.cmake"
+        "${CMAKE_CURRENT_BINARY_DIR}/BaseQISSeleneInterfaceConfig.cmake"
     DESTINATION
-        ${CMAKE_INSTALL_LIBDIR}/cmake/HeliosSeleneInterface
+        ${CMAKE_INSTALL_LIBDIR}/cmake/BaseQISSeleneInterface
 )
