+++ selene-ext/interfaces/base_qis/c/cmake/BaseQISSeleneInterfaceConfig.cmake.in
@@ -1,0 +1,9 @@
+@PACKAGE_INIT@
+
+include(CMakeFindDependencyMacro)
+
+find_dependency(Selene REQUIRED)
+
+include("${CMAKE_CURRENT_LIST_DIR}/BaseQISSeleneInterfaceTargets.cmake")
+
+check_required_components(BaseQISSeleneInterface)
