+++ selene-ext/interfaces/base_qis/c/include/base_qis/random.h
@@ -1,0 +1,13 @@
+#ifndef HELIOS_QIS_RANDOM_H
+#define HELIOS_QIS_RANDOM_H
+
+#include <stdint.h>
+#include <base_qis/macros.h>
+
+EXPORT void random_seed(uint64_t seed);
+EXPORT void random_advance(uint64_t delta);
+EXPORT uint32_t random_int();
+EXPORT uint32_t random_rng(uint32_t bound);
+EXPORT double random_float();
+
+#endif
