+++ selene-ext/interfaces/base_qis/c/include/base_qis/selene_lifetime.h
@@ -1,0 +1,9 @@
+#ifndef HELIOS_QIS_SELENE_LIFETIME_H
+#define HELIOS_QIS_SELENE_LIFETIME_H
+
+#include <selene/selene.h>
+#include <base_qis/macros.h>
+
+extern SeleneInstance* selene_instance;
+
+#endif
