+++ selene-ext/interfaces/base_qis/c/src/selene_lifetime.c
@@ -1,0 +1,3 @@
+#include <base_qis/selene_lifetime.h>
+
+SeleneInstance* selene_instance = 0;
