+++ selene-ext/interfaces/helios_qis/c/cmake/HeliosSeleneInterfaceConfig.cmake.in
@@ -1,5 +1,10 @@
 @PACKAGE_INIT@
 
-include("${CMAKE_CURRENT_LIST_DIR}/SeleneTargets.cmake")
+include(CMakeFindDependencyMacro)
 
-set(Selene_INCLUDE_DIRS "${PACKAGE_PREFIX_DIR}/include")
+find_dependency(Selene REQUIRED)
+find_dependency(BaseQISSeleneInterface REQUIRED)
+
+include("${CMAKE_CURRENT_LIST_DIR}/HeliosSeleneInterfaceTargets.cmake")
+
+check_required_components(HeliosSeleneInterface)
