+++ selene-ext/interfaces/helios_qis/python/selene_helios_qis_plugin/__init__.py
@@ -1,3 +1,4 @@
-from .plugin import HeliosInterface, LogLevel
+from .plugin import HeliosInterface
+from selene_base_qis_plugin import LogLevel
 
 __all__ = ["HeliosInterface", "LogLevel"]
