+++ selene-ext/utilities/argreader/build.rs
@@ -2,23 +2,24 @@ use std::env;
 use std::path::PathBuf;
 
 fn main() {
-    println!("cargo:rerun-if-env-changed=SELENE_HELIOS_QIS_LIB_DIR");
+    println!("cargo:rerun-if-env-changed=SELENE_BASE_QIS_LIB_DIR");
 
-    let lib_dir = env::var_os("SELENE_HELIOS_QIS_LIB_DIR")
+    let lib_dir = env::var_os("SELENE_BASE_QIS_LIB_DIR")
         .map(PathBuf::from)
-        .unwrap_or_else(default_helios_lib_dir);
+        .unwrap_or_else(default_base_lib_dir);
     if !lib_dir.exists() {
         panic!(
-            "Helios interface runtime directory not found: {}. Build the Helios QIS interface first or set SELENE_HELIOS_QIS_LIB_DIR.",
+            "Helios interface runtime directory not found: {}. Build the Helios QIS interface first or set SELENE_BASE_QIS_LIB_DIR.",
             lib_dir.display()
         );
     }
 
     println!("cargo:rustc-link-search=native={}", lib_dir.display());
-    println!("cargo:rustc-link-lib=dylib=helios_selene_interface_runtime");
+    println!("cargo:rustc-link-lib=dylib=base_qis_selene_interface");
+    println!("cargo:rustc-link-arg=-Wl,-rpath,$ORIGIN/../../../selene_base_qis_plugin/_dist/lib")
 }
 
-fn default_helios_lib_dir() -> PathBuf {
+fn default_base_lib_dir() -> PathBuf {
     PathBuf::from(env!("CARGO_MANIFEST_DIR"))
-        .join("../../interfaces/helios_qis/python/selene_helios_qis_plugin/_dist/lib")
+        .join("../../interfaces/base_qis/python/selene_base_qis_plugin/_dist/lib")
 }
