+++ selene-sim/cbindgen.toml
@@ -14,7 +14,7 @@ language = "C"
 
 # header = "/* Text to put at the beginning of the generated file. Probably a license. */"
 # trailer = "/* Text to put at the end of the generated file */"
-# include_guard = "my_bindings_h"
+include_guard = "SELENE_SIM_H"
 # pragma_once = true
 # autogen_warning = "/* Warning, this file is autogenerated by cbindgen. Don't modify this manually. */"
 include_version = false
