+++ selene-sim/python/selene_sim/build.py
@@ -63,10 +63,10 @@ def _collect_libdeps(
     from selene_helios_qis_plugin import HeliosInterface
 
     interface = interface or HeliosInterface()
-    deps = [LibDep.from_plugin(interface)]
+    deps = LibDep.from_plugin(interface)
     interface.register_build_steps(planner)
     for u in utilities or []:
-        deps.append(LibDep.from_plugin(u))
+        deps.extend(LibDep.from_plugin(u))
     return deps
